refactor lang info form for cache key
