+++ frontend/talentsearch/src/js/components/languageInformationForm/LanguageInformationForm.tsx
@@ -1,8 +1,13 @@
 import React, { ReactNode } from "react";
 import { useIntl } from "react-intl";
 import { errorMessages, navigationMessages } from "@common/messages";
-import { Checklist, RadioGroup, Select } from "@common/components/form";
-import { FormProvider, SubmitHandler, useForm } from "react-hook-form";
+import {
+  BasicForm,
+  Checklist,
+  RadioGroup,
+  Select,
+} from "@common/components/form";
+import { SubmitHandler } from "react-hook-form";
 import { enumToOptions } from "@common/helpers/formUtils";
 import compact from "lodash/compact";
 import omit from "lodash/omit";
@@ -27,6 +32,7 @@ import applicantProfileRoutes from "../../applicantProfileRoutes";
 import directIntakeRoutes from "../../directIntakeRoutes";
 import profileMessages from "../profile/profileMessages";
 import getFullPoolAdvertisementTitle from "../pool/getFullPoolAdvertisementTitle";
+import ConsideredLanguages from "./ConsideredLanguages";
 
 export type FormValues = Pick<
   User,
@@ -110,13 +116,7 @@ export const LanguageInformationForm: React.FunctionComponent<{
       ? directIntakePaths.reviewApplication(application.id)
       : profilePaths.home(initialData.id);
 
-  const defaultValues = dataToFormValues(initialData);
-
-  const methods = useForm({
-    defaultValues,
-  });
-
-  const onSubmit: SubmitHandler<FormValues> = async (formValues) => {
+  const handleSubmit: SubmitHandler<FormValues> = async (formValues) => {
     await submitHandler(initialData.id, formValuesToSubmitData(formValues))
       .then(() => {
         navigate(returnRoute);
@@ -125,48 +125,6 @@ export const LanguageInformationForm: React.FunctionComponent<{
       .catch(() => {
         toast.error(intl.formatMessage(profileMessages.updatingFailed));
       });
-  };
-
-  // hooks to watch, needed for conditional rendering
-  const [consideredLanguages, bilingualEvaluation] = methods.watch([
-    "consideredPositionLanguages",
-    "bilingualEvaluation",
-  ]);
-
-  const languageEvaluationPageLink = () => {
-    return (
-      <a
-        target="_blank"
-        rel="noopener noreferrer"
-        href={
-          locale === "en"
-            ? "https://www.canada.ca/en/public-service-commission/services/second-language-testing-public-service.html"
-            : "https://www.canada.ca/fr/commission-fonction-publique/services/evaluation-langue-seconde.html"
-        }
-      >
-        {intl.formatMessage({
-          defaultMessage: "Government of Canada language evaluation.",
-          id: "Ugr5Yt",
-          description: "Message on links to the language evaluation tests",
-        })}
-      </a>
-    );
-  };
-
-  const selfAssessmentLink = (msg: React.ReactNode): React.ReactNode => {
-    return (
-      <a
-        target="_blank"
-        rel="noopener noreferrer"
-        href={
-          locale === "en"
-            ? "https://www.canada.ca/en/public-service-commission/services/second-language-testing-public-service/self-assessment-tests.html"
-            : "https://www.canada.ca/fr/commission-fonction-publique/services/evaluation-langue-seconde/tests-autoevaluation.html"
-        }
-      >
-        {msg}
-      </a>
-    );
   };
 
   const ConsideredLangItems: { value: string; label: string }[] = [
@@ -195,100 +153,6 @@ export const LanguageInformationForm: React.FunctionComponent<{
       }),
     },
   ];
-  const BilingualEvaluationItems: {
-    value: string;
-    label: string | ReactNode;
-  }[] = [
-    {
-      value: BilingualEvaluation.CompletedEnglish,
-      label: intl.formatMessage(
-        {
-          defaultMessage:
-            "I am bilingual (En/Fr) and <strong>have</strong> completed an official <strong>ENGLISH</strong> <languageEvaluationPageLink></languageEvaluationPageLink>",
-          id: "ljnSYf",
-          description:
-            "Message for the completed english bilingual evaluation option",
-        },
-        {
-          languageEvaluationPageLink,
-        },
-      ),
-    },
-    {
-      value: BilingualEvaluation.CompletedFrench,
-      label: intl.formatMessage(
-        {
-          defaultMessage:
-            "I am bilingual (En/Fr) and <strong>have</strong> completed an official <strong>FRENCH</strong> <languageEvaluationPageLink></languageEvaluationPageLink>",
-          id: "pak8ye",
-          description:
-            "Message for the completed french bilingual evaluation option",
-        },
-        {
-          languageEvaluationPageLink,
-        },
-      ),
-    },
-    {
-      value: BilingualEvaluation.NotCompleted,
-      label: intl.formatMessage(
-        {
-          defaultMessage:
-            "I am bilingual (En/Fr) and <strong>have NOT</strong> completed an official <languageEvaluationPageLink></languageEvaluationPageLink>",
-          id: "FfEyFv",
-          description:
-            "Message for the haven't completed bilingual evaluation option",
-        },
-        {
-          languageEvaluationPageLink,
-        },
-      ),
-    },
-  ];
-  const EvaluatedAbilityItemsSortOrder = [
-    EvaluatedLanguageAbility.X,
-    EvaluatedLanguageAbility.A,
-    EvaluatedLanguageAbility.B,
-    EvaluatedLanguageAbility.C,
-    EvaluatedLanguageAbility.E,
-    EvaluatedLanguageAbility.P,
-  ];
-  const EvaluatedAbilityItems: { value: string; label: string }[] =
-    enumToOptions(EvaluatedLanguageAbility, EvaluatedAbilityItemsSortOrder).map(
-      ({ value }) => ({
-        value,
-        label: value,
-      }),
-    );
-  const EstimatedAbilityItems: { value: string; label: string | ReactNode }[] =
-    [
-      {
-        value: EstimatedLanguageAbility.Beginner,
-        label: intl.formatMessage({
-          defaultMessage:
-            "Beginner <gray>- I have basic reading, writing and verbal communication skills.</gray>",
-          id: "ZuFBx5",
-          description: "Message for the beginner language ability option",
-        }),
-      },
-      {
-        value: EstimatedLanguageAbility.Intermediate,
-        label: intl.formatMessage({
-          defaultMessage:
-            "Intermediate <gray>- I have strong reading, writing and verbal communication skills.</gray>",
-          id: "t5G3Fz",
-          description: "Message for the intermediate language ability option",
-        }),
-      },
-      {
-        value: EstimatedLanguageAbility.Advanced,
-        label: intl.formatMessage({
-          defaultMessage: "Advanced <gray>- I am completely fluent.</gray>",
-          id: "paLFgh",
-          description: "Message for the advanced language ability option",
-        }),
-      },
-    ];
 
   const applicationBreadcrumbs = application
     ? [
@@ -317,202 +181,72 @@ export const LanguageInformationForm: React.FunctionComponent<{
     : [];
 
   return (
-    <FormProvider {...methods}>
-      <form onSubmit={methods.handleSubmit(onSubmit)}>
-        <ProfileFormWrapper
-          description={intl.formatMessage({
-            defaultMessage:
-              "Use the form below to help us better understand your language preferences and capabilities",
-            id: "TGCq/w",
-            description:
-              "Description text for Profile Form wrapper in Language Information Form",
-          })}
-          title={intl.formatMessage({
-            defaultMessage: "Language Information",
-            id: "R5aTZ9",
-            description:
-              "Title for Profile Form wrapper in Language Information Form",
-          })}
-          cancelLink={{
-            href: returnRoute,
-          }}
-          crumbs={[
-            ...applicationBreadcrumbs,
-            {
-              title: intl.formatMessage({
-                defaultMessage: "Language Information",
-                id: "/k21MP",
+    <BasicForm
+      cacheKey="lang-info-form"
+      onSubmit={handleSubmit}
+      options={{
+        defaultValues: dataToFormValues(initialData),
+      }}
+    >
+      <ProfileFormWrapper
+        description={intl.formatMessage({
+          defaultMessage:
+            "Use the form below to help us better understand your language preferences and capabilities",
+          id: "TGCq/w",
+          description:
+            "Description text for Profile Form wrapper in Language Information Form",
+        })}
+        title={intl.formatMessage({
+          defaultMessage: "Language Information",
+          id: "R5aTZ9",
+          description:
+            "Title for Profile Form wrapper in Language Information Form",
+        })}
+        cancelLink={{
+          href: returnRoute,
+        }}
+        crumbs={[
+          ...applicationBreadcrumbs,
+          {
+            title: intl.formatMessage({
+              defaultMessage: "Language Information",
+              id: "/k21MP",
+              description:
+                "Display Text for Language Information Form Page Link",
+            }),
+          },
+        ]}
+        prefixBreadcrumbs={!application}
+      >
+        <div data-h2-padding="base(0, 0, x2, 0)">
+          <div
+            data-h2-width="base(100%) p-tablet(75%) l-tablet(50%)"
+            data-h2-padding="base(x.5, 0)"
+          >
+            <Checklist
+              idPrefix="considered-position-languages"
+              legend={intl.formatMessage({
+                defaultMessage:
+                  "Select the positions you would like to be considered for",
+                id: "ntUOoz",
                 description:
-                  "Display Text for Language Information Form Page Link",
-              }),
-            },
-          ]}
-          prefixBreadcrumbs={!application}
-        >
-          <div data-h2-padding="base(0, 0, x2, 0)">
-            <div
-              data-h2-width="base(100%) p-tablet(75%) l-tablet(50%)"
-              data-h2-padding="base(x.5, 0)"
-            >
-              <Checklist
-                idPrefix="considered-position-languages"
-                legend={intl.formatMessage({
-                  defaultMessage:
-                    "Select the positions you would like to be considered for",
-                  id: "ntUOoz",
-                  description:
-                    "Legend for considered position languages check list in language information form",
-                })}
-                name="consideredPositionLanguages"
-                rules={{
-                  required: intl.formatMessage(errorMessages.required),
-                }}
-                items={ConsideredLangItems}
-              />
-            </div>
-            {consideredLanguages.includes("lookingForBilingual") && (
-              <>
-                <div data-h2-padding="base(x.5, 0, 0, 0)">
-                  <RadioGroup
-                    idPrefix="bilingualEvaluation"
-                    legend={intl.formatMessage({
-                      defaultMessage: "Bilingual evaluation",
-                      id: "X354at",
-                      description:
-                        "Legend bilingual evaluation status in language information form",
-                    })}
-                    name="bilingualEvaluation"
-                    rules={{
-                      required: intl.formatMessage(errorMessages.required),
-                    }}
-                    items={BilingualEvaluationItems}
-                  />
-                </div>
-                {bilingualEvaluation !== BilingualEvaluation.NotCompleted ? (
-                  <div data-h2-padding="base(x.5, 0, 0, 0)">
-                    <p>
-                      {intl.formatMessage({
-                        defaultMessage:
-                          "Please indicate the language levels you acquired from your Government of Canada language evaluation.",
-                        id: "Y7l4/n",
-                        description:
-                          "Text requesting language levels given from bilingual evaluation in language information form",
-                      })}
-                    </p>
-                    <div data-h2-flex-grid="base(normal, x1)">
-                      <div data-h2-flex-item="base(1of1) p-tablet(1of3) desktop(1of4)">
-                        <Select
-                          id="comprehensionLevel"
-                          name="comprehensionLevel"
-                          label={intl.formatMessage({
-                            defaultMessage: "Comprehension",
-                            id: "W4Svkd",
-                            description:
-                              "Label displayed on the language information form comprehension field.",
-                          })}
-                          nullSelection={intl.formatMessage({
-                            defaultMessage: "Select a level...",
-                            id: "8QN6ZC",
-                            description:
-                              "Placeholder displayed on the language information form comprehension field.",
-                          })}
-                          rules={{
-                            required: intl.formatMessage(
-                              errorMessages.required,
-                            ),
-                          }}
-                          options={EvaluatedAbilityItems}
-                        />
-                      </div>
-                      <div data-h2-flex-item="base(1of1) p-tablet(1of3) desktop(1of4)">
-                        <Select
-                          id="writtenLevel"
-                          name="writtenLevel"
-                          label={intl.formatMessage({
-                            defaultMessage: "Written",
-                            id: "x5C9Ab",
-                            description:
-                              "Label displayed on the language information form written field.",
-                          })}
-                          nullSelection={intl.formatMessage({
-                            defaultMessage: "Select a level...",
-                            id: "aQJOd0",
-                            description:
-                              "Placeholder displayed on the language information form written field.",
-                          })}
-                          rules={{
-                            required: intl.formatMessage(
-                              errorMessages.required,
-                            ),
-                          }}
-                          options={EvaluatedAbilityItems}
-                        />
-                      </div>
-                      <div data-h2-flex-item="base(1of1) p-tablet(1of3) desktop(1of4)">
-                        <Select
-                          id="verbalLevel"
-                          name="verbalLevel"
-                          label={intl.formatMessage({
-                            defaultMessage: "Verbal",
-                            id: "rywI3C",
-                            description:
-                              "Label displayed on the language information form verbal field.",
-                          })}
-                          nullSelection={intl.formatMessage({
-                            defaultMessage: "Select a level...",
-                            id: "Y7jEXr",
-                            description:
-                              "Placeholder displayed on the language information form verbal field.",
-                          })}
-                          rules={{
-                            required: intl.formatMessage(
-                              errorMessages.required,
-                            ),
-                          }}
-                          options={EvaluatedAbilityItems}
-                        />
-                      </div>
-                    </div>
-                  </div>
-                ) : (
-                  <div data-h2-padding="base(x.5, 0, 0, 0)">
-                    <p data-h2-padding="base(0, 0, x.5, 0)">
-                      {intl.formatMessage(
-                        {
-                          defaultMessage:
-                            "If you want to find out your language proficiency levels, <selfAssessmentLink>click here to find out.</selfAssessmentLink>",
-                          id: "nVh2Qh",
-                          description:
-                            "Text including link to language proficiency evaluation in language information form",
-                        },
-                        {
-                          selfAssessmentLink,
-                        },
-                      )}
-                    </p>
-                    <RadioGroup
-                      idPrefix="estimatedLanguageAbility"
-                      legend={intl.formatMessage({
-                        defaultMessage: "Second language proficiency level",
-                        id: "T1TKNL",
-                        description:
-                          "Legend for second language proficiency level in language information form",
-                      })}
-                      name="estimatedLanguageAbility"
-                      items={EstimatedAbilityItems}
-                    />
-                  </div>
-                )}
-              </>
-            )}
+                  "Legend for considered position languages check list in language information form",
+              })}
+              name="consideredPositionLanguages"
+              rules={{
+                required: intl.formatMessage(errorMessages.required),
+              }}
+              items={ConsideredLangItems}
+            />
           </div>
-          <ProfileFormFooter
-            mode="saveButton"
-            cancelLink={{ href: returnRoute }}
-          />
-        </ProfileFormWrapper>
-      </form>
-    </FormProvider>
+          <ConsideredLanguages />
+        </div>
+        <ProfileFormFooter
+          mode="saveButton"
+          cancelLink={{ href: returnRoute }}
+        />
+      </ProfileFormWrapper>
+    </BasicForm>
   );
 };
